Rat in a maze
If one rat starts moving from start vertex to destination vertex, we have to find that is there any way to complete the path, if it is possible then mark the correct path for the rat. The maze is given using a binary matrix, where it is marked with 1, it is a valid path, otherwise 0 for a blocked cell.